Started on writing closure AST as code
