+++ src/intrinsics/mod.rs
@@ -2,9 +2,7 @@
 
 use std::collections::HashMap;
 use std::sync::{Arc, Mutex};
-use std::io;
 use rand::Rng;
-use piston_meta::json;
 
 use runtime::{Flow, Runtime, Side};
 use ast;
@@ -370,141 +368,6 @@ pub fn standard(f: &mut Prelude) {
     });
 }
 
-enum EscapeString {
-    Json,
-    None
-}
-
-
-fn write_variable<W>(
-    w: &mut W,
-    rt: &Runtime,
-    v: &Variable,
-    escape_string: EscapeString
-) -> Result<(), io::Error>
-    where W: io::Write
-{
-    match *v {
-        Variable::Text(ref t) => {
-            match escape_string {
-                EscapeString::Json => {
-                    try!(json::write_string(w, t));
-                }
-                EscapeString::None => {
-                    try!(write!(w, "{}", t))
-                }
-            }
-        }
-        Variable::F64(x, _) => {
-            try!(write!(w, "{}", x));
-        }
-        Variable::Vec4(v) => {
-            try!(write!(w, "({}, {}", v[0], v[1]));
-            if v[2] != 0.0 || v[3] != 0.0 {
-                try!(write!(w, ", {}", v[2]));
-                if v[3] != 0.0 {
-                    try!(write!(w, ", {})", v[3]));
-                } else {
-                    try!(write!(w, ")"));
-                }
-            } else {
-                try!(write!(w, ")"));
-            }
-        }
-        Variable::Bool(x, _) => {
-            try!(write!(w, "{}", x));
-        }
-        Variable::Ref(ind) => {
-            try!(write_variable(w, rt, &rt.stack[ind], escape_string));
-        }
-        Variable::Link(ref link) => {
-            match escape_string {
-                EscapeString::Json => {
-                    // Write link items.
-                    try!(write!(w, "link {{ "));
-                    for slice in &link.slices {
-                        for i in slice.start..slice.end {
-                            let v = slice.block.var(i);
-                            try!(write_variable(w, rt, &v, EscapeString::Json));
-                            try!(write!(w, " "));
-                        }
-                    }
-                    try!(write!(w, "}}"));
-                }
-                EscapeString::None => {
-                    for slice in &link.slices {
-                        for i in slice.start..slice.end {
-                            let v = slice.block.var(i);
-                            try!(write_variable(w, rt, &v, EscapeString::None));
-                        }
-                    }
-                }
-            }
-        }
-        Variable::Object(ref obj) => {
-            try!(write!(w, "{{"));
-            let n = obj.len();
-            for (i, (k, v)) in obj.iter().enumerate() {
-                try!(write!(w, "{}: ", k));
-                try!(write_variable(w, rt, v, EscapeString::Json));
-                if i + 1 < n {
-                    try!(write!(w, ", "));
-                }
-            }
-            try!(write!(w, "}}"));
-        }
-        Variable::Array(ref arr) => {
-            try!(write!(w, "["));
-            let n = arr.len();
-            for (i, v) in arr.iter().enumerate() {
-                try!(write_variable(w, rt, v, EscapeString::Json));
-                if i + 1 < n {
-                    try!(write!(w, ", "));
-                }
-            }
-            try!(write!(w, "]"));
-        }
-        Variable::Option(ref opt) => {
-            match opt {
-                &None => {
-                    try!(write!(w, "none()"))
-                }
-                &Some(ref v) => {
-                    try!(write!(w, "some("));
-                    try!(write_variable(w, rt, v, EscapeString::Json));
-                    try!(write!(w, ")"));
-                }
-            }
-        }
-        Variable::Result(ref res) => {
-            match res {
-                &Err(ref err) => {
-                    try!(write!(w, "err("));
-                    try!(write_variable(w, rt, &err.message,
-                                        EscapeString::Json));
-                    try!(write!(w, ")"));
-                }
-                &Ok(ref ok) => {
-                    try!(write!(w, "ok("));
-                    try!(write_variable(w, rt, ok, EscapeString::Json));
-                    try!(write!(w, ")"));
-                }
-            }
-        }
-        Variable::Thread(_) => try!(write!(w, "_thread")),
-        Variable::Return => try!(write!(w, "_return")),
-        Variable::UnsafeRef(_) => try!(write!(w, "_unsafe_ref")),
-        Variable::RustObject(_) => try!(write!(w, "_rust_object")),
-        Variable::Closure(_, _) => try!(write!(w, "_closure")),
-        // ref x => panic!("Could not print out `{:?}`", x)
-    }
-    Ok(())
-}
-
-fn print_variable(rt: &Runtime, v: &Variable, escape_string: EscapeString) {
-    write_variable(&mut io::stdout(), rt, v, escape_string).unwrap();
-}
-
 pub fn call_standard(
     rt: &mut Runtime,
     index: usize,
@@ -769,6 +632,8 @@ fn println(
     lc: usize,
     cu: usize,
 ) -> Result<Option<Variable>, String> {
+    use write::{print_variable, EscapeString};
+
     rt.push_fn(call.name.clone(), 0, None, st, lc, cu);
     let x = rt.stack.pop().expect(TINVOTS);
     print_variable(rt, &x, EscapeString::None);
@@ -785,6 +650,8 @@ fn print(
     lc: usize,
     cu: usize,
 ) -> Result<Option<Variable>, String> {
+    use write::{print_variable, EscapeString};
+
     rt.push_fn(call.name.clone(), 0, None, st, lc, cu);
     let x = rt.stack.pop().expect(TINVOTS);
     print_variable(rt, &x, EscapeString::None);
@@ -1427,6 +1294,8 @@ fn _str(
     lc: usize,
     cu: usize,
 ) -> Result<Option<Variable>, String> {
+    use write::{write_variable, EscapeString};
+
     rt.push_fn(call.name.clone(), 0, None, st + 1, lc, cu);
     let v = rt.stack.pop().expect(TINVOTS);
     let mut buf: Vec<u8> = vec![];
@@ -1444,6 +1313,8 @@ fn json_string(
     _lc: usize,
     _cu: usize,
 ) -> Result<Option<Variable>, String> {
+    use write::{write_variable, EscapeString};
+
     let v = rt.stack.pop().expect(TINVOTS);
     let mut buf: Vec<u8> = vec![];
     write_variable(&mut buf, rt, rt.resolve(&v), EscapeString::Json).unwrap();
@@ -2148,6 +2019,8 @@ fn unwrap(
     lc: usize,
     cu: usize,
 ) -> Result<Option<Variable>, String> {
+    use write::{write_variable, EscapeString};
+
     // Return value does not depend on lifetime of argument since
     // `ok(x)` and `some(x)` perform a deep clone.
     rt.push_fn(call.name.clone(), 0, None, st + 1, lc, cu);
@@ -2458,6 +2331,7 @@ fn save__data_file(
 ) -> Result<Option<Variable>, String> {
     use std::error::Error;
     use std::fs::File;
+    use write::{write_variable, EscapeString};
 
     rt.push_fn(call.name.clone(), 0, None, st + 1, lc, cu);
     let file = rt.stack.pop().expect(TINVOTS);
