+++ src/lib.rs
@@ -22,6 +22,7 @@ pub mod ty;
 pub mod link;
 pub mod macros;
 pub mod vec4;
+pub mod write;
 
 mod grab;
